CE-Ecosystem Builder/requirements.txt
scrapegraphai
ollama
pandas
geopy
pydantic
beautifulsoup4
networkx
pyyaml
tqdm
playwright
requests
httpx  # Async HTTP client for faster I/O operations
